Created abstract class to handle a bridge between an OpenMrsObject and is position in the form.
Extended conditions and obs from this new class.

updated liquibase-update-to-latest-2.4.x.xml to add the new column form_namespace_path to the table conditions.
