+++ api/src/main/java/org/openmrs/Obs.java
@@ -61,7 +61,7 @@ import org.slf4j.LoggerFactory;
  * 
  * @see Encounter
  */
-public class Obs extends BaseChangeableOpenmrsData {
+public class Obs extends FormRecordable{
 	
 	/**
 	 * @since 2.1.0
@@ -86,10 +86,6 @@ public class Obs extends BaseChangeableOpenmrsData {
 	public static final long serialVersionUID = 112342333L;
 	
 	private static final Logger log = LoggerFactory.getLogger(Obs.class);
-	
-	private static final String FORM_NAMESPACE_PATH_SEPARATOR = "^";
-	
-	private static final int FORM_NAMESPACE_PATH_MAX_LENGTH = 255;
 	
 	protected Integer obsId;
 	
@@ -855,7 +851,7 @@ public class Obs extends BaseChangeableOpenmrsData {
 	 * Get the value for the ComplexData. This method is used by the ComplexObsHandler. The
 	 * valueComplex has two parts separated by a bar '|' character: part A) the title; and part B)
 	 * the URI. The title is the readable description of the valueComplex that is returned by
-	 * {@link Obs#getValueAsString()}. The URI is the location where the ComplexData is stored.
+	 * {@link Obs#getValueAsString(Locale)}. The URI is the location where the ComplexData is stored.
 	 * 
 	 * @return readable title and URI for the location of the ComplexData binary object.
 	 * @since 1.5
@@ -1177,94 +1173,6 @@ public class Obs extends BaseChangeableOpenmrsData {
 	}
 	
 	/**
-	 * Gets the namespace for the form field that was used to capture the obs details in the form
-	 * 
-	 * @return the namespace
-	 * @since 1.11
-	 * <strong>Should</strong> return the namespace for a form field that has no path
-	 * <strong>Should</strong> return the correct namespace for a form field with a path
-	 * <strong>Should</strong> return null if the namespace is not specified
-	 */
-	public String getFormFieldNamespace() {
-		if (StringUtils.isNotBlank(formNamespaceAndPath)) {
-			//Only the path was specified
-			if (formNamespaceAndPath.startsWith(FORM_NAMESPACE_PATH_SEPARATOR)) {
-				return null;
-			}
-			return formNamespaceAndPath.substring(0, formNamespaceAndPath.indexOf(FORM_NAMESPACE_PATH_SEPARATOR));
-		}
-		
-		return formNamespaceAndPath;
-	}
-	
-	/**
-	 * Gets the path for the form field that was used to capture the obs details in the form
-	 * 
-	 * @return the the form field path
-	 * @since 1.11
-	 * <strong>Should</strong> return the path for a form field that has no namespace
-	 * <strong>Should</strong> return the correct path for a form field with a namespace
-	 * <strong>Should</strong> return null if the path is not specified
-	 */
-	public String getFormFieldPath() {
-		if (StringUtils.isNotBlank(formNamespaceAndPath)) {
-			//Only the namespace was specified
-			if (formNamespaceAndPath.endsWith(FORM_NAMESPACE_PATH_SEPARATOR)) {
-				return null;
-			}
-			return formNamespaceAndPath.substring(formNamespaceAndPath.indexOf(FORM_NAMESPACE_PATH_SEPARATOR) + 1);
-		}
-		
-		return formNamespaceAndPath;
-	}
-	
-	/**
-	 * Sets the namespace and path of the form field that was used to capture the obs details in the
-	 * form.<br>
-	 * <b>Note:</b> Namespace and formFieldPath together must not exceed 254 characters in length,
-	 * form applications can subtract the length of their namespace from 254 to determine the
-	 * maximum length they can use for a form field path.
-	 * 
-	 * @param namespace the namespace of the form field
-	 * @param formFieldPath the path of the form field
-	 * @since 1.11
-	 * <strong>Should</strong> set the underlying formNamespaceAndPath in the correct pattern
-	 * <strong>Should</strong> reject a namepace containing the separator
-	 * <strong>Should</strong> reject a path containing the separator
-	 * <strong>Should</strong> reject a namepace and path combination longer than the max length
-	 * <strong>Should</strong> not mark the obs as dirty when the value has not been changed
-	 * <strong>Should</strong> mark the obs as dirty when the value has been changed
-	 * <strong>Should</strong> mark the obs as dirty when the value is changed from a null to a non null value
-	 * <strong>Should</strong> mark the obs as dirty when the value is changed from a non null to a null value
-	 */
-	public void setFormField(String namespace, String formFieldPath) {
-		if (namespace == null && formFieldPath == null) {
-			markAsDirty(formNamespaceAndPath, null);
-			formNamespaceAndPath = null;
-			return;
-		}
-		
-		String nsAndPathTemp = "";
-		if (StringUtils.isNotBlank(namespace) && StringUtils.isNotBlank(formFieldPath)) {
-			nsAndPathTemp = namespace + FORM_NAMESPACE_PATH_SEPARATOR + formFieldPath;
-		} else if (StringUtils.isNotBlank(namespace)) {
-			nsAndPathTemp = namespace + FORM_NAMESPACE_PATH_SEPARATOR;
-		} else if (StringUtils.isNotBlank(formFieldPath)) {
-			nsAndPathTemp = FORM_NAMESPACE_PATH_SEPARATOR + formFieldPath;
-		}
-		
-		if (nsAndPathTemp.length() > FORM_NAMESPACE_PATH_MAX_LENGTH) {
-			throw new APIException("Obs.namespaceAndPathTooLong", (Object[]) null);
-		}
-		if (StringUtils.countMatches(nsAndPathTemp, FORM_NAMESPACE_PATH_SEPARATOR) > 1) {
-			throw new APIException("Obs.namespaceAndPathNotContainSeparator", (Object[]) null);
-		}
-		
-		markAsDirty(this.formNamespaceAndPath, nsAndPathTemp);
-		formNamespaceAndPath = nsAndPathTemp;
-	}
-	
-	/**
 	 * Returns true if any change has been made to an Obs instance. In general, the only time
 	 * isDirty() is going to return false is when a new Obs has just been instantiated or loaded
 	 * from the database and no method that modifies it internally has been invoked.
@@ -1281,7 +1189,8 @@ public class Obs extends BaseChangeableOpenmrsData {
 		return dirty;
 	}
 	
-	private void markAsDirty(Object oldValue, Object newValue) {
+	@Override
+	protected void markAsDirty(Object oldValue, Object newValue) {
 		//Should we ignore the case for Strings?
 		if (!isDirty() && obsId != null && !OpenmrsUtil.nullSafeEquals(oldValue, newValue)) {
 			dirty = true;
